# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Micro para gestión de usuarios, clientes y técnicos",
        "title": "Microservicio de usuario UDC",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/user-service/v1",
    "paths": {
...
        "/changePassword": {
            "patch": {
                "description": "Cambia la contraseña de un técnico existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "Cambiar contraseña",
                "parameters": [
                    {
                        "description": "Datos para cambiar contraseña",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChangePasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ejemplo: {\\\"message\\\":\\\"PASSWORD_CHANGED_SUCCESSFULLY\\\"}",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/createTechnician": {
            "post": {
                "description": "Crea un nuevo técnico en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "Crear técnico",
                "parameters": [
                    {
                        "description": "Datos del técnico",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTechnicianDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TechnicianDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/deleteTechnician/{technicianID}": {
            "delete": {
                "description": "Elimina un técnico por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "Eliminar técnico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del técnico",
                        "name": "technicianID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ejemplo: {\\\"message\\\":\\\"TECHNICIAN_DELETED_SUCCESSFULLY\\\"}",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/technician/{technicianID}": {
            "get": {
                "description": "Busca un técnico por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "Obtener técnico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del técnico",
                        "name": "technicianID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TechnicianDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/updateTechnician/{technicianID}": {
            "put": {
                "description": "Actualiza los datos de un técnico existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "Actualizar técnico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del técnico",
                        "name": "technicianID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del técnico",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTechnicianDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TechnicianDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/common.ErrorResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ChangePasswordDTO": {
            "type": "object",
            "required": [
                "NewPassword",
                "UserId"
            ],
            "properties": {
                "NewPassword": {
                    "type": "string",
                    "example": "newSecret123"
                },
                "UserId": {
                    "type": "string",
                    "example": "carlos.r"
                }
            }
        },
...
        "dto.CreateTechnicianDTO": {
            "type": "object"
        },
...
        "dto.TechnicianDTO": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string",
                    "example": "Calle 123 #45-67"
                },
                "Document": {
                    "type": "string",
                    "example": "1234567890"
                },
                "DocumentType": {
                    "type": "string",
                    "example": "CC"
                },
                "Email": {
                    "type": "string",
                    "example": "carlos.ramirez@example.com"
                },
                "EntryDate": {
                    "type": "string",
                    "example": "2025-01-01T15:04:05Z"
                },
                "ID": {
                    "type": "string",
                    "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "IsActive": {
                    "type": "boolean",
                    "example": true
                },
                "Name": {
                    "type": "string",
                    "example": "Carlos"
                },
                "Password": {
                    "type": "string",
                    "example": "$2a$10$hashPassword"
                },
                "PhoneNumber": {
                    "type": "string",
                    "example": "3001234567"
                },
                "Surname": {
                    "type": "string",
                    "example": "Ramírez"
                },
                "Username": {
                    "type": "string",
                    "example": "carlos.r"
                }
            }
        },
...
        "dto.UpdateTechnicianDTO": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string",
                    "example": "Carrera 45 #12-34"
                },
                "Document": {
                    "type": "string",
                    "example": "9876543210"
                },
                "DocumentTypeId": {
                    "type": "string",
                    "example": "CC"
                },
                "Email": {
                    "type": "string",
                    "example": "juan.perez@example.com"
                },
                "Name": {
                    "type": "string",
                    "example": "Juan"
                },
                "PhoneNumber": {
                    "type": "string",
                    "example": "3017654321"
                }
            }
        },
...
    }
}